Squash the last 4 commits after deleting the >100MB file
The APK file was causing problems because it was too large. It
was removed and the repository needed to move 4 commits back
to remove all the cached information and be able to keep working
forward on it.

The generate_tfrecord script was added to create the TFRecord files
from the CSV files.

Also and the README was updated.